+++ docs/_about/changelog.md
@@ -9,6 +9,39 @@ New versions are released through GitHub, so the reference page is the [GitHub R
 > Starting from 2.4.0, you can now [support development](https://github.com/sponsors/natario1) through the GitHub Sponsors program. 
 Companies can share a tiny part of their revenue and get private support hours in return. Thanks!
 
+##### v2.7.2
+
+- Fix: fix camera rotation handling for Compose apps and other specific scenarios ([#1117][1117])
+
+<https://github.com/natario1/CameraView/compare/v2.7.1...v2.7.2>
+
+##### v2.7.1
+
+- Fix: fix preview issues on Pixel 4A with certain FPS, thanks to [@honzasmuk][honzasmuk] ([#1089][1089])
+- Improvement: don't catch gestures if they're turned off, thanks to [@ObsidianX][ObsidianX] ([#1068][1068])
+- New: new flag cameraDrawHardwareOverlays and setDrawHardwareOverlays() to draw overlays on hardware canvas, thanks to [@ObsidianX][ObsidianX] ([#1066][1066])
+- Publish on Maven Central instead of JCenter
+
+<https://github.com/natario1/CameraView/compare/v2.7.0...v2.7.1>
+
+##### v2.7.0
+
+- New: onPictureShutter() callback when taking pictures, thanks to [@EzequielAdrianM][EzequielAdrianM] ([#1030][1030])
+- New: GestureAction.TAKE_PICTURE_SNAPSHOT lets you take snapshots on gesture, thanks to [@EzequielAdrianM][EzequielAdrianM] ([#1030][1030])
+- Improvement: try-catch internal exception when takePicture fails, thanks to [@michaelspecht][michaelspecht] ([#1024][1024])
+- Improvement: log errors when file writing fails, thanks to [@bwt][bwt] ([#960][960])
+- Fix: Avoid preview deadlocks ([#1020][1020])
+- Fix: Workaround for messed-up preview on Pixel 4 ([#1020][1020])
+- Fix: Avoid internal StackOverflow errors ([#992][992])
+
+<https://github.com/natario1/CameraView/compare/v2.6.4...v2.7.0>
+
+##### v2.6.4
+
+- Fix: Fix many small bugs ([#953][953])
+
+<https://github.com/natario1/CameraView/compare/v2.6.3...v2.6.4>
+
 ##### v2.6.3
 
 - <small>[Video]</small> New: `setAudioCodec` and `app:cameraAudioCodec` to choose the audio encoding format, thanks to [@EverydayPineapple][EverydayPineapple] ([#861][861])
@@ -316,6 +349,7 @@ This is the last release before v2.
 
 [aartikov]: https://github.com/aartikov
 [athornz]: https://github.com/athornz
+[bwt]: https://github.com/bwt
 [v-gar]: https://github.com/v-gar
 [andrewmunn]: https://github.com/andrewmunn
 [chaitanyaraghav]: https://github.com/chaitanyaraghav
@@ -338,6 +372,10 @@ This is the last release before v2.
 [hualong-shen]: https://github.com/hualong-shen
 [EverydayPineapple]: https://github.com/EverydayPineapple
 [jeffreyfjohnson]: https://github.com/jeffreyfjohnson
+[michaelspecht]: https://github.com/michaelspecht
+[EzequielAdrianM]: https://github.com/EzequielAdrianM
+[honzasmuk]: https://github.com/honzasmuk
+[ObsidianX]: https://github.com/ObsidianX
 
 
 [73]: https://github.com/natario1/CameraView/pull/73
@@ -435,3 +473,13 @@ This is the last release before v2.
 [861]: https://github.com/natario1/CameraView/pull/861
 [877]: https://github.com/natario1/CameraView/pull/877
 [897]: https://github.com/natario1/CameraView/pull/897
+[953]: https://github.com/natario1/CameraView/pull/953
+[960]: https://github.com/natario1/CameraView/pull/960
+[992]: https://github.com/natario1/CameraView/pull/992
+[1020]: https://github.com/natario1/CameraView/pull/1020
+[1024]: https://github.com/natario1/CameraView/pull/1024
+[1030]: https://github.com/natario1/CameraView/pull/1030
+[1089]: https://github.com/natario1/CameraView/pull/1089
+[1068]: https://github.com/natario1/CameraView/pull/1068
+[1066]: https://github.com/natario1/CameraView/pull/1066
+[1117]: https://github.com/natario1/CameraView/pull/1117